Bunch of perestroika here:
Removed the 'bin' prefix on many php files, updated references to those files to reflect this.
Started paving the way for the node service to be the primary interface for device configuration.
Cleaned up stray 'demo' pages and checked their contents.
Removed the PHP system interface (such that it was) and consolidated its code in etc.php.
Tweaked the help pane code and styles.
Removed extraneous pages.
Moved relevant pages to obtaining their configuration from etc.js if they did not already and removed the PHP backend pages previously used for this purpose.